+++ src/pages/App.js
@@ -4,12 +4,13 @@ import Roster from './Roster'
 import Awards from './Awards'
 import Ranks from './Ranks'
 import Qualifications from './Qualifications'
-import { Error } from '../components/Error'
 import { Footer } from '../components/Footer'
 import { config } from '../constants'
 import User from './User'
 import Forms from './Forms'
 import Form from './Form'
+import Calendar from './Calendar'
+import { Alert } from '../components/Alert'
 
 function App() {
   const [searchParams] = useSearchParams()
@@ -25,6 +26,7 @@ function App() {
           <Route path='/awards' element={<Awards />}></Route>
           <Route path='/forms' element={<Forms />}></Route>
           <Route path='/forms/:id' element={<Form />}></Route>
+          <Route path='/calendar' element={<Calendar />}></Route>
           <Route path='/qualifications' element={<Qualifications />}></Route>
           <Route path='/ranks' element={<Ranks />}></Route>
           <Route path='/roster' element={<Roster />}></Route>
@@ -32,7 +34,7 @@ function App() {
           <Route path='*' element={<Navigate to={`/${useLocation().search}`} />}></Route>
         </Routes>
       ) : (
-        <Error error='Please make sure all required parameters have been included.' />
+        <Alert message='Please make sure all required widget parameters have been included.' type='danger' />
       )}
       <Footer />
     </div>